@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Attributes List
-class AttributeList {
+class AttributeList implements Model {
     /// Total number of attributes in the given collection.
     final int total;
     /// List of attributes.
