@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Buckets List
-class BucketList {
+class BucketList implements Model {
     /// Total number of buckets documents that matched your query.
     final int total;
     /// List of buckets.
