@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Currency
-class Currency {
+class Currency implements Model {
     /// Currency symbol.
     final String symbol;
     /// Currency name.
