@@ -1,21 +1,29 @@
 part of dart_appwrite.models;
 
 /// Database
-class Database {
+class Database implements Model {
     /// Database ID.
     final String $id;
     /// Database name.
     final String name;
+    /// Database creation date in ISO 8601 format.
+    final String $createdAt;
+    /// Database update date in ISO 8601 format.
+    final String $updatedAt;
 
     Database({
         required this.$id,
         required this.name,
+        required this.$createdAt,
+        required this.$updatedAt,
     });
 
     factory Database.fromMap(Map<String, dynamic> map) {
         return Database(
             $id: map['\$id'].toString(),
             name: map['name'].toString(),
+            $createdAt: map['\$createdAt'].toString(),
+            $updatedAt: map['\$updatedAt'].toString(),
         );
     }
 
@@ -23,6 +31,8 @@ class Database {
         return {
             "\$id": $id,
             "name": name,
+            "\$createdAt": $createdAt,
+            "\$updatedAt": $updatedAt,
         };
     }
 }
