@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Databases List
-class DatabaseList {
+class DatabaseList implements Model {
     /// Total number of databases documents that matched your query.
     final int total;
     /// List of databases.
