@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Deployments List
-class DeploymentList {
+class DeploymentList implements Model {
     /// Total number of deployments documents that matched your query.
     final int total;
     /// List of deployments.
