@@ -1,15 +1,15 @@
 part of dart_appwrite.models;
 
 /// Execution
-class Execution {
+class Execution implements Model {
     /// Execution ID.
     final String $id;
-    /// Execution creation date in Unix timestamp.
-    final int $createdAt;
-    /// Execution update date in Unix timestamp.
-    final int $updatedAt;
-    /// Execution read permissions.
-    final List $read;
+    /// Execution creation date in ISO 8601 format.
+    final String $createdAt;
+    /// Execution upate date in ISO 8601 format.
+    final String $updatedAt;
+    /// Execution roles.
+    final List $permissions;
     /// Function ID.
     final String functionId;
     /// The trigger that caused the function to execute. Possible values can be: `http`, `schedule`, or `event`.
@@ -20,7 +20,9 @@ class Execution {
     final int statusCode;
     /// The script response output string. Logs the last 4,000 characters of the execution response output.
     final String response;
-    /// The script stderr output string. Logs the last 4,000 characters of the execution stderr output
+    /// The script stdout output string. Logs the last 4,000 characters of the execution stdout output. This will return an empty string unless the response is returned using an API key or as part of a webhook payload.
+    final String stdout;
+    /// The script stderr output string. Logs the last 4,000 characters of the execution stderr output. This will return an empty string unless the response is returned using an API key or as part of a webhook payload.
     final String stderr;
     /// The script execution time in seconds.
     final double time;
@@ -29,12 +31,13 @@ class Execution {
         required this.$id,
         required this.$createdAt,
         required this.$updatedAt,
-        required this.$read,
+        required this.$permissions,
         required this.functionId,
         required this.trigger,
         required this.status,
         required this.statusCode,
         required this.response,
+        required this.stdout,
         required this.stderr,
         required this.time,
     });
@@ -42,14 +45,15 @@ class Execution {
     factory Execution.fromMap(Map<String, dynamic> map) {
         return Execution(
             $id: map['\$id'].toString(),
-            $createdAt: map['\$createdAt'],
-            $updatedAt: map['\$updatedAt'],
-            $read: map['\$read'],
+            $createdAt: map['\$createdAt'].toString(),
+            $updatedAt: map['\$updatedAt'].toString(),
+            $permissions: map['\$permissions'],
             functionId: map['functionId'].toString(),
             trigger: map['trigger'].toString(),
             status: map['status'].toString(),
             statusCode: map['statusCode'],
             response: map['response'].toString(),
+            stdout: map['stdout'].toString(),
             stderr: map['stderr'].toString(),
             time: map['time'].toDouble(),
         );
@@ -60,12 +64,13 @@ class Execution {
             "\$id": $id,
             "\$createdAt": $createdAt,
             "\$updatedAt": $updatedAt,
-            "\$read": $read,
+            "\$permissions": $permissions,
             "functionId": functionId,
             "trigger": trigger,
             "status": status,
             "statusCode": statusCode,
             "response": response,
+            "stdout": stdout,
             "stderr": stderr,
             "time": time,
         };
