@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Executions List
-class ExecutionList {
+class ExecutionList implements Model {
     /// Total number of executions documents that matched your query.
     final int total;
     /// List of executions.
