@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Files List
-class FileList {
+class FileList implements Model {
     /// Total number of files documents that matched your query.
     final int total;
     /// List of files.
