@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Functions List
-class FunctionList {
+class FunctionList implements Model {
     /// Total number of functions documents that matched your query.
     final int total;
     /// List of functions.
