@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Health Queue
-class HealthQueue {
+class HealthQueue implements Model {
     /// Amount of actions in the queue.
     final int size;
 
