@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Health Status
-class HealthStatus {
+class HealthStatus implements Model {
     /// Duration in milliseconds how long the health check took.
     final int ping;
     /// Service status. Possible values can are: `pass`, `fail`
