@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Indexes List
-class IndexList {
+class IndexList implements Model {
     /// Total number of indexes documents that matched your query.
     final int total;
     /// List of indexes.
