@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Language
-class Language {
+class Language implements Model {
     /// Language name.
     final String name;
     /// Language two-character ISO 639-1 codes.
