@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Logs List
-class LogList {
+class LogList implements Model {
     /// Total number of logs documents that matched your query.
     final int total;
     /// List of logs.
