@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Memberships List
-class MembershipList {
+class MembershipList implements Model {
     /// Total number of memberships documents that matched your query.
     final int total;
     /// List of memberships.
