@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Phones List
-class PhoneList {
+class PhoneList implements Model {
     /// Total number of phones documents that matched your query.
     final int total;
     /// List of phones.
