@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Runtimes List
-class RuntimeList {
+class RuntimeList implements Model {
     /// Total number of runtimes documents that matched your query.
     final int total;
     /// List of runtimes.
