@@ -1,23 +1,23 @@
 part of dart_appwrite.models;
 
 /// Session
-class Session {
+class Session implements Model {
     /// Session ID.
     final String $id;
-    /// Session creation date in Unix timestamp.
-    final int $createdAt;
+    /// Session creation date in ISO 8601 format.
+    final String $createdAt;
     /// User ID.
     final String userId;
-    /// Session expiration date in Unix timestamp.
-    final int expire;
+    /// Session expiration date in ISO 8601 format.
+    final String expire;
     /// Session Provider.
     final String provider;
     /// Session Provider User ID.
     final String providerUid;
     /// Session Provider Access Token.
     final String providerAccessToken;
-    /// Date, the Unix timestamp of when the access token expires.
-    final int providerAccessTokenExpiry;
+    /// The date of when the access token expires in ISO 8601 format.
+    final String providerAccessTokenExpiry;
     /// Session Provider Refresh Token.
     final String providerRefreshToken;
     /// IP in use when the session was created.
@@ -84,13 +84,13 @@ class Session {
     factory Session.fromMap(Map<String, dynamic> map) {
         return Session(
             $id: map['\$id'].toString(),
-            $createdAt: map['\$createdAt'],
+            $createdAt: map['\$createdAt'].toString(),
             userId: map['userId'].toString(),
-            expire: map['expire'],
+            expire: map['expire'].toString(),
             provider: map['provider'].toString(),
             providerUid: map['providerUid'].toString(),
             providerAccessToken: map['providerAccessToken'].toString(),
-            providerAccessTokenExpiry: map['providerAccessTokenExpiry'],
+            providerAccessTokenExpiry: map['providerAccessTokenExpiry'].toString(),
             providerRefreshToken: map['providerRefreshToken'].toString(),
             ip: map['ip'].toString(),
             osCode: map['osCode'].toString(),
