@@ -1,13 +1,13 @@
 part of dart_appwrite.models;
 
 /// Team
-class Team {
+class Team implements Model {
     /// Team ID.
     final String $id;
-    /// Team creation date in Unix timestamp.
-    final int $createdAt;
-    /// Team update date in Unix timestamp.
-    final int $updatedAt;
+    /// Team creation date in ISO 8601 format.
+    final String $createdAt;
+    /// Team update date in ISO 8601 format.
+    final String $updatedAt;
     /// Team name.
     final String name;
     /// Total number of team members.
@@ -24,8 +24,8 @@ class Team {
     factory Team.fromMap(Map<String, dynamic> map) {
         return Team(
             $id: map['\$id'].toString(),
-            $createdAt: map['\$createdAt'],
-            $updatedAt: map['\$updatedAt'],
+            $createdAt: map['\$createdAt'].toString(),
+            $updatedAt: map['\$updatedAt'].toString(),
             name: map['name'].toString(),
             total: map['total'],
         );
