@@ -1,17 +1,17 @@
 part of dart_appwrite.models;
 
 /// Token
-class Token {
+class Token implements Model {
     /// Token ID.
     final String $id;
-    /// Token creation date in Unix timestamp.
-    final int $createdAt;
+    /// Token creation date in ISO 8601 format.
+    final String $createdAt;
     /// User ID.
     final String userId;
     /// Token secret key. This will return an empty string unless the response is returned using an API key or as part of a webhook payload.
     final String secret;
-    /// Token expiration date in Unix timestamp.
-    final int expire;
+    /// Token expiration date in ISO 8601 format.
+    final String expire;
 
     Token({
         required this.$id,
@@ -24,10 +24,10 @@ class Token {
     factory Token.fromMap(Map<String, dynamic> map) {
         return Token(
             $id: map['\$id'].toString(),
-            $createdAt: map['\$createdAt'],
+            $createdAt: map['\$createdAt'].toString(),
             userId: map['userId'].toString(),
             secret: map['secret'].toString(),
-            expire: map['expire'],
+            expire: map['expire'].toString(),
         );
     }
 
