@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Users List
-class UserList {
+class UserList implements Model {
     /// Total number of users documents that matched your query.
     final int total;
     /// List of users.
